openssl: link system openssl.conf after installation
In the compilation of a package which needs `openssl.conf` there is a
crash due this file is missing in the paths specified by Spack, e.g.,:
`+Can't open
/opt/spack/opt/spack/linux-fedora34-skylake/gcc-7.5.0/openssl-1.1.1l-bnwhtmqzbgx5cke7cc4nvkrpqt6wyr6k/etc/openssl/openssl.cnf
for reading, No such file or directory`. The added lines fix this
issue by a symlink to the system `openssl` configuration file.